RavenDB-13111 Moving common functionality to helpers/utils
